Lib

Documentation

partMaybes :: [Maybe a] -> [Maybe [Maybe a]]

The "natural" partition of a list of Maybe's: we group Nothing's and Just's and the outermost Maybe marks which is contained.

unMaybeParts :: [Maybe [Maybe a]] -> [[Maybe a]]

Convert the "natural" partition into a partition that we can more easily push through the innermost Maybe: replace Nothing's with an empty lists and drop Just's.

partMaybe :: [Maybe a] -> [[Maybe a]]

unMaybeParts . partMaybes

dartMaybe :: [[Maybe a]] -> [Maybe [a]]

Reassociate the list and Maybe

pushMaybe :: [Maybe a] -> [Maybe [a]]

dartMaybe . partMaybe

fromPushed :: [Maybe [a]] -> [Maybe a]

Convert from pushed preserving Nothing's

simplifyNothings :: [Maybe a] -> [Maybe a]

Replace runs of multiple Nothing's with single Nothing's

fromPushedIsSimplified :: Eq a => [Maybe a] -> Bool

fromPushed . pushMaybe == simplifyNothings

flattenPushed :: [Maybe [a]] -> [a]

Drop the Nothing's and flatten the resulting list

flattenPushedIsCatMaybes :: Eq a => [Maybe a] -> Bool

flattenPushed . pushMaybe == catMaybes

toBin :: (Bits a, Eq a, Enum a) => a -> [a]

To a reversed list of he binary digits of the input

toMaybes :: [Bool] -> [a] -> [Maybe a]

If False then Nothing else Just, zipped

enumMaybes :: (Bits a, Eq a, Enum a) => Int -> a -> [Maybe a]

Use toBin to get the list of Bool's and use [1..] as the list of values. The first argument is the length and the second is the index.

testEnumMaybes :: ([Maybe Int] -> Bool) -> Int -> Int -> [[Maybe Int]]

Apply a test to enumMaybes up to the given maximums. Return any counterexamples.

assertEnumMaybes :: String -> ([Maybe Int] -> Bool) -> Int -> Int -> IO ()

Throw an error if the test fails on any input enumMaybes

runTests :: IO ()

digraphNotes :: ()

Notes on digraphs:

 newtype Digraph = Digraph { digraphMatrix :: Matrix Bool }

 data DigraphGen = DigraphGen
   { currentDigraph :: Digraph
   , currentArc :: Arc
   , newDigraph :: Bool
   }

 We can enumerate the non (i, i) arcs in (join liftM2 (,) [1..n])

 this gives us an Enum instance for (ArcN (n :: Nat))

 newtype ArcN (n :: Nat) = ArcN { toArc :: Arc }